refactor: move methods that modify parameter values to a separate class (#862)
