+++ samtranslator/public/sdk/parameter.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+
+from samtranslator.sdk.parameter import SamParameterValues
